# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_eleven-am_voice-backend_internal_dto.AgentListResponse:
    properties:
      agents:
        items:
          $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_dto.AgentResponse'
        type: array
    type: object
  github_com_eleven-am_voice-backend_internal_dto.AgentResponse:
    properties:
      active_installs:
        example: 500
        type: integer
      avg_rating:
        example: 4.5
        type: number
      capabilities:
        example:
        - search
        - weather
        items:
          type: string
        type: array
      category:
        example: productivity
        type: string
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      description:
        example: A helpful voice assistant
        type: string
      developer_id:
        example: usr_xyz789
        type: string
      id:
        example: agt_abc123
        type: string
      is_public:
        example: true
        type: boolean
      is_verified:
        example: false
        type: boolean
      keywords:
        example:
        - helpful
        - assistant
        items:
          type: string
        type: array
      logo_url:
        example: https://example.com/logo.png
        type: string
      name:
        example: My Voice Agent
        type: string
      total_installs:
        example: 1000
        type: integer
      total_reviews:
        example: 100
        type: integer
      updated_at:
        example: "2024-01-20T15:45:00Z"
        type: string
    type: object
...
  github_com_eleven-am_voice-backend_internal_dto.CreateAgentRequest:
    properties:
      capabilities:
        example:
        - search
        - weather
        - calendar
        items:
          type: string
        type: array
      category:
        enum:
        - productivity
        - education
        - entertainment
        - lifestyle
        - utility
        - social
        - finance
        - health
        - news
        - assistant
        example: productivity
        type: string
      description:
        example: A helpful voice assistant
        type: string
      keywords:
        example:
        - helpful
        - assistant
        - voice
        items:
          type: string
        type: array
      logo_url:
        example: https://example.com/logo.png
        type: string
      name:
        example: My Voice Agent
        type: string
    type: object
...
  github_com_eleven-am_voice-backend_internal_dto.InstallRequest:
    properties:
      scopes:
        example:
        - profile
        - email
        - realtime
        items:
          type: string
        type: array
    type: object
  github_com_eleven-am_voice-backend_internal_dto.InstalledAgentResponse:
    properties:
      agent_id:
        example: agt_abc123
        type: string
      description:
        example: A helpful voice assistant
        type: string
      granted_scopes:
        example:
        - profile
        - email
        items:
          type: string
        type: array
      installed_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      logo_url:
        example: https://example.com/logo.png
        type: string
      name:
        example: My Voice Agent
        type: string
    type: object
  github_com_eleven-am_voice-backend_internal_dto.InstalledAgentsResponse:
    properties:
      agents:
        items:
          $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_dto.InstalledAgentResponse'
        type: array
    type: object
  github_com_eleven-am_voice-backend_internal_dto.MarketplaceAgentResponse:
    properties:
      avg_rating:
        example: 4.8
        type: number
      category:
        example: productivity
        type: string
      description:
        example: A helpful voice assistant
        type: string
      id:
        example: agt_abc123
        type: string
      is_verified:
        example: true
        type: boolean
      keywords:
        example:
        - helpful
        - assistant
        items:
          type: string
        type: array
      logo_url:
        example: https://example.com/logo.png
        type: string
      name:
        example: My Voice Agent
        type: string
      total_installs:
        example: 5000
        type: integer
      total_reviews:
        example: 200
        type: integer
    type: object
  github_com_eleven-am_voice-backend_internal_dto.MarketplaceListResponse:
    properties:
      agents:
        items:
          $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_dto.MarketplaceAgentResponse'
        type: array
      limit:
        example: 20
        type: integer
      offset:
        example: 0
        type: integer
    type: object
...
  github_com_eleven-am_voice-backend_internal_dto.UpdateAgentRequest:
    properties:
      capabilities:
        example:
        - new
        - capabilities
        items:
          type: string
        type: array
      category:
        enum:
        - productivity
        - education
        - entertainment
        - lifestyle
        - utility
        - social
        - finance
        - health
        - news
        - assistant
        example: education
        type: string
      description:
        example: Updated description
        type: string
      keywords:
        example:
        - updated
        - keywords
        items:
          type: string
        type: array
      logo_url:
        example: https://example.com/new-logo.png
        type: string
      name:
        example: Updated Agent Name
        type: string
    type: object
  github_com_eleven-am_voice-backend_internal_dto.UpdateScopesRequest:
    properties:
      scopes:
        example:
        - profile
        - email
        items:
          type: string
        type: array
    type: object
  github_com_eleven-am_voice-backend_internal_shared.APIError:
    properties:
      code:
        example: invalid_request
        type: string
      details:
        type: object
      message:
        example: Invalid request body
        type: string
    type: object
host: api.voice.example.com
info:
  contact: {}
  description: API server for voice agent platform
  title: Voice Backend API
  version: 1.0.0
paths:
  /agents:
    get:
      description: Returns all agents owned by the authenticated developer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_dto.AgentListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
      summary: List developer's agents
      tags:
      - agents
    post:
      consumes:
      - application/json
      description: Creates a new agent for the authenticated developer
      parameters:
      - description: Agent details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_dto.CreateAgentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_dto.AgentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
      summary: Create a new agent
      tags:
      - agents
  /agents/{id}:
    delete:
      description: Deletes an agent owned by the authenticated developer
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
      summary: Delete an agent
      tags:
      - agents
    get:
      description: Returns a specific agent owned by the authenticated developer
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_dto.AgentResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
      summary: Get an agent
      tags:
      - agents
    put:
      consumes:
      - application/json
      description: Updates an existing agent owned by the authenticated developer
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      - description: Agent update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_dto.UpdateAgentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_dto.AgentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
      summary: Update an agent
      tags:
      - agents
...
  /me/agents:
    get:
      description: Returns all agents installed by the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_dto.InstalledAgentsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
      summary: List installed agents
      tags:
      - installs
  /me/agents/{id}:
    delete:
      description: Removes an installed agent from the user's account
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
      summary: Uninstall an agent
      tags:
      - installs
  /me/agents/{id}/install:
    post:
      consumes:
      - application/json
      description: Installs a public agent for the authenticated user
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      - description: Installation options
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_dto.InstallRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_dto.InstalledAgentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
      summary: Install an agent
      tags:
      - installs
  /me/agents/{id}/scopes:
    put:
      consumes:
      - application/json
      description: Updates the granted scopes for an installed agent
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      - description: New scopes
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_dto.UpdateScopesRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
      summary: Update agent scopes
      tags:
      - installs
...
  /store/agents:
    get:
      description: Returns paginated list of public agents in the marketplace
      parameters:
      - description: Number of results (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      - description: Filter by category
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_dto.MarketplaceListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
      summary: List public agents
      tags:
      - marketplace
  /store/agents/{id}:
    get:
      description: Returns details of a public agent by ID
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_dto.MarketplaceAgentResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eleven-am_voice-backend_internal_shared.APIError'
      summary: Get a public agent
      tags:
      - marketplace
...
securityDefinitions:
  APIKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  SessionAuth:
    in: cookie
    name: session
    type: apiKey
swagger: "2.0"